Move rendering tests to the render module
